pyproject.toml
[project]
name = "fastapi-beyond-crud"
version = "1.0.0"
description = "Simple api with Fastapi"

authors = [
    {name = "teigorme", email = "igorteixeira354@gmail.com"},
]
dependencies = ["fastapi[standard]>=0.117.1", "sqlmodel>=0.0.25", "asyncpg>=0.30.0", "alembic>=1.16.5", "pydantic-settings>=2.11.0", "passlib>=1.7.4"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "fastapi run src/main.py"
dev = "fastapi dev src/main.py"
alembic = "alembic init -t async migrations"
